Extract shard transfer helpers into module
